+++ src/avx2_seeding.rs
@@ -8,8 +8,8 @@ use crate::types::*;
 pub unsafe fn mm_hash256(kmer: __m256i) -> __m256i {
     let mut key = kmer;
     let s1 = _mm256_slli_epi64(key, 21);
-    key = _mm256_add_epi64(key, s1);
     key = _mm256_xor_si256(key, _mm256_cmpeq_epi64(key, key));
+    key = _mm256_add_epi64(key, s1);
 
     key = _mm256_xor_si256(key, _mm256_srli_epi64(key, 24));
     let s2 = _mm256_slli_epi64(key, 3);